requirements.txt
airtable==0.3.1
alembic==1.4.0
amqp==2.5.2
bcrypt==3.1.7
billiard==3.6.2.0
blinker==1.4
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
event-bus==1.0.2
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jwt==0.6.1
kombu==4.6.7
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-enum==1.5.1
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
sentry-sdk==0.14.1
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.0
zipp==2.2.0
